0.0.0.15 - не корректен ГО

+++ app/domain/value_objects/BreakdownItem.py
@@ -6,11 +6,11 @@ class BreakdownItem:
     description: str  # Описание строки
     volume: float     # Объем для расчета
     percent: float    # Процент
-    money: float      # Деньги
+    money: int      # Деньги
 
 @dataclass
 class IncomeBreakdown:
     personal_items: List[BreakdownItem]      # Личные начисления
     group_items: List[BreakdownItem]         # Групповые начисления
     leader_items: List[BreakdownItem]        # Лидерские начисления
-    total_money: float                       # Итоговая сумма
+    total_money: int                       # Итоговая сумма
