+++ app/web/scheme/schemas.py
@@ -22,23 +22,23 @@ class BranchInfo(BaseModel):
     branch_id: int
     branch_qualification_by_gv: str  # Квалификация по group volume
     branch_qualification_by_side: str  # Квалификация по side volume
-    lo: float
-    gv: float
-    side_volume: float
+    lo: int
+    gv: int
+    side_volume: int
     is_closed: bool  # Закрыла ли квалификацию по side volume
     is_stronger_than_parent: bool  # Сильнее ли родителя
-    parent_earn_percent: float  # Сколько % получает родитель с этой ветки
-    parent_earn_money: float  # Сколько денег получает родитель с этой ветки
-    level: float
+    parent_earn_percent: int  # Сколько % получает родитель с этой ветки
+    parent_earn_money: int  # Сколько денег получает родитель с этой ветки
+    level: int
 
 
 class IncomeResponse(BaseModel):
     user_id: int
     qualification: str
-    lo: float
-    go: float
-    side_volume: float
-    points: float
+    lo: int
+    go: int
+    side_volume: int
+    points: int
     personal_bonus: float
     structure_bonus: float
     mentor_bonus: float
@@ -48,19 +48,19 @@ class IncomeResponse(BaseModel):
     leader_money: int
     total_money: int
     veron: int
-    total_income: float
+    total_income: int
     branches_info: List[BranchInfo]  # НОВОЕ ПОЛЕ
 
 
 class AddLORequest(BaseModel):
-    lo: float
+    lo: int
 
 
 class MemberStatus(BaseModel):
     user_id: int
-    lo: float
-    go: float
-    level: float
+    lo: int
+    go: int
+    level: int
     personal_bonus: float
     structure_bonus: float
     total_income: float
